Eliminate errors in mes_strings.c

+++ mes.h
@@ -77,7 +77,7 @@ struct scm
 struct scm *g_cells;
 struct scm *g_news;
 SCM vector_entry(SCM x);
-long MAX_STRING;
+int MAX_STRING;
 char *g_buf;
 long g_free;
 int g_debug;
